@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class BaseApiController < ApplicationController
-  before_filter :parse_request, :authenticate_user_from_token!
+  before_action :parse_request, :authenticate_user_from_token!
 
   private
 
