@@ -2,4 +2,4 @@
 
 # Be sure to restart your server when you modify this file.
 
-Rails.application.config.session_store :cookie_store, key: '_secretsantaapp_session'
+Rails.application.config.session_store :cookie_store, key: '_secret_santa_app_session'
